[AC-5666] move functional_expertise_type to types directory.
