+++ web/impact/impact/graphql/types/expert_profile_type.py
@@ -6,8 +6,9 @@ from impact.graphql.types.industry_type import IndustryType  # noqa: F401
 from impact.graphql.types.startup_type import StartupType
 from impact.graphql.types.program_family_type import ProgramFamilyType  # noqa: F401, E501
 from impact.graphql.types.user_type import UserType  # noqa: F401
-from impact.graphql.functional_expertise_type import FunctionalExpertiseType  # noqa: F401, E501
-
+from impact.graphql.types.functional_expertise_type import (
+    FunctionalExpertiseType  # noqa: F401, E501
+)
 
 class ExpertProfileType(DjangoObjectType):
     mentees = graphene.List(StartupType)
