+++ web/impact/impact/graphql/types/functional_expertise_type.py
